remove github_repos_contributors
