+++ backend/common/infrastructure/src/database/schema.rs
@@ -223,13 +223,6 @@ diesel::table! {
 }
 
 diesel::table! {
-    github_repos_contributors (repo_id, user_id) {
-        repo_id -> Int8,
-        user_id -> Int8,
-    }
-}
-
-diesel::table! {
     github_user_indexes (user_id) {
         user_id -> Int8,
         user_indexer_state -> Nullable<Jsonb>,
@@ -425,7 +418,6 @@ diesel::allow_tables_to_appear_in_same_query!(
     github_pull_requests,
     github_repo_indexes,
     github_repos,
-    github_repos_contributors,
     github_user_indexes,
     github_users,
     ignored_github_issues,
