+++ backend/event-listeners/src/listeners/mod.rs
@@ -74,7 +74,6 @@ pub async fn spawn_all(
 			database.clone(),
 			database.clone(),
 			database.clone(),
-			database.clone(),
 		)
 		.spawn(
 			event_bus::consumer_with_exchange(
