+++ hasura/metadata/databases/default/tables/tables.yaml
@@ -20,7 +20,6 @@
 - "!include public_github_issues.yaml"
 - "!include public_github_pull_requests.yaml"
 - "!include public_github_repos.yaml"
-- "!include public_github_repos_contributors.yaml"
 - "!include public_github_users.yaml"
 - "!include public_ignored_github_issues.yaml"
 - "!include public_onboardings.yaml"
